requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.4
django-allauth==0.43.0
django-crispy-forms==1.10.0
django-debug-toolbar==3.2
django-on-heroku==1.1.2
django-rq==2.4.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.4
redis==3.5.3
requests==2.25.0
requests-oauthlib==1.3.0
rq==1.7.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
